readme links added and such
added links to the other sections as well as made the section about how I decided on the keymap collapsed so people don't have to read it unless they wanted to. Need to add how to use in the read me file or the other files like icons.h and such.