Added how to section in readme file
Added a how to use section that runs through flashing the board, however I realize now I didn't include the hex file so I need to add that later. I added links that hopefully work, as well as some pics to the flashing section, I need to finish the compiling section as well as add more links to other sections, and add photos.